Acceleration Bands

acceleration_bands() is a generic S3 function that preserves the input class: data.frame in, data.frame out; matrix in, matrix out.

Handling of -values

Leading NAs are always produced for the initial lookback period where insufficient data is available. If the input itself contains NAs they are passed through to the underlying C routine, which can cause the entire output to be filled with NAs. Set na.ignore = TRUE to strip NAs before calculation and re-insert them at their original positions in the output.

Usage

acceleration_bands(x, cols, n = 10, na.ignore = FALSE, ...)

Arguments

x

An OHLC-V series coercible to data.frame.

cols

(formula). An optional 3-variable formula selecting columns from x via model.frame. Defaults to ~high + low + close.

n

(integer). Lookback period (window size). A positive integer of length 1.

na.ignore

(logical). A logical of length 1. FALSE by default. If TRUE, NAs in the input are stripped before calculation and re-inserted at their original positions in the output.

...

Additional parameters passed into model.frame

Value

An object of same class and length of x:

UpperBand

double

MiddleBand

double

LowerBand

double

See also

Author

Serkan Korkmaz

Examples

## load Bitcoin (BTC)
## series
data(BTC, package = "talib")

## calculate the indicator
## for Bitcoin (BTC)
output <- talib::acceleration_bands(BTC)

## display the results
utils::tail(output)
#>                     UpperBand MiddleBand LowerBand
#> 2024-12-26 01:00:00  110264.9   98217.88  87213.56
#> 2024-12-27 01:00:00  109387.1   97020.16  85808.60
#> 2024-12-28 01:00:00  107244.3   96516.01  86184.86
#> 2024-12-29 01:00:00  105457.8   96134.41  86850.78
#> 2024-12-30 01:00:00  104630.5   95620.42  87225.80
#> 2024-12-31 01:00:00  104522.1   95236.42  86566.28

## visualize the indicator
## with talib::chart()
##
## see ?talib::chart or ?talib::indicator
## for more details
{
 ## chart OHLC-V
 ## series with talib::chart()
 talib::chart(BTC)

 ## chart indicator
 ## with default values
 talib::indicator(
     talib::acceleration_bands
 )
}