Average Directional Movement Index Rating

The average_directional_movement_index_rating() is a generic S3 function that builds upon 'type-safe'-esque workflows limited to classes in in base R, and the package-wide dependencies. Ie. class in, class out. Each method is a soft wrapper of model.frame and therefore the OHLC-V series must be coercible to a data.frame.

Handling of -values

average_directional_movement_index_rating() iterates over valid values, and returns NA for the remaing part of series.

Usage

average_directional_movement_index_rating(
  x,
  cols,
  n = 10,
  na.ignore = FALSE,
  ...
)

Arguments

x

An OHLC-V series that is coercible to data.frame.

cols

(formula). An optional 3 variable formula passed into model.frame. Internally uses ~high + low + close by default.

n

(integer). An integer of length 1.

na.ignore

(logical). A logical of length 1. FALSE by default. If TRUE 's are ignored during calculation to avoid returning x filled with 's.

...

Additional parameters passed into model.frame

Value

An object of same class and length of x:

ADXR

double

See also

Author

Serkan Korkmaz

Examples

## load Bitcoin (BTC)
## series
data(BTC, package = "talib")

## calculate the indicator
## for Bitcoin (BTC)
output <- talib::average_directional_movement_index_rating(BTC)

## display the results
utils::tail(output)
#>                         ADXR
#> 2024-12-26 01:00:00 34.62534
#> 2024-12-27 01:00:00 33.27814
#> 2024-12-28 01:00:00 31.97181
#> 2024-12-29 01:00:00 31.69110
#> 2024-12-30 01:00:00 31.33726
#> 2024-12-31 01:00:00 30.93191

## visualize the indicator
## with talib::chart()
##
## see ?talib::chart or ?talib::indicator
## for more details
{
 ## chart OHLC-V
 ## series with talib::chart()
 talib::chart(BTC)

 ## chart indicator
 ## with default values
 talib::indicator(
     talib::average_directional_movement_index_rating
 )
}