Bollinger Bands

The bollinger_bands() is a generic S3 function that builds upon 'type-safe'-esque workflows limited to classes in in base R, and the package-wide dependencies. Ie. class in, class out. Each method is a soft wrapper of model.frame and therefore the OHLC-V series must be coercible to a data.frame. This rule does not transfer to indicators that uses univariate series, unless passed as a 1-column data.frame or matrix. In such cases, if univariate series is passed as a vector the function calculates the indicator 'as is', and returns a data.frame if the indicator itself is also a univariate series.

The indicator, by default, follows its mathematical definition. However, the cols argument allows for simple rearrangement of the definition by passing relevant columns in a custom order. Refer to the details-section for more on the calculation of the indicators.

Usage

bollinger_bands(x, cols, ma = SMA(n = 10), up = 2, down = 2, ...)

Arguments

x

An OHLC-V series that is coercible to data.frame. The function assumes that all columns are named in lowercase and order invariant.

cols

An optional formula passed into model.frame. If passed into indicators based on univariate series, the function calculates indicators for each element in 'cols'. For indicators based on multivariate series, it will alter the calculation itself. See vignette("talib") for more details.

ma

A function call to a moving average function.

up, down

A pair of double for upper and lower standard deviations.

...

Additional parameters passed into model.frame

Value

A data.frame- or matrix-object with the format:

upper

double. The lower band.

middle

double. The middle band.

lower

double. The upper band.

See also

Author

Serkan Korkmaz

Examples

## load Bitcoin (BTC)
## series
data(BTC, package = "talib")

## calculate the indicator
## for Bitcoin (BTC)
output <- talib::bollinger_bands(BTC)

## display the results
utils::tail(output)
#>         upper   middle    lower
#> 361 106812.14 99260.81 91709.48
#> 362 104471.11 98218.27 91965.44
#> 363 100874.66 97021.36 93168.05
#> 364  99891.29 96519.09 93146.89
#> 365  99875.94 96136.93 92397.91
#> 366  99720.50 95622.57 91524.65

## visualize the indicator
## with candlesticks
##
## see ?talib::chart or ?talib::indicator
## for more details
{
 ## chart OHLC-V
 ## series with candlesticks
 talib::chart(BTC)

 ## chart indicator
 ## with default values
 talib::indicator(
     talib::bollinger_bands
 )
}