Hilbert Transform - Phasor Components

The ht_phasor() is a generic S3 function that builds upon 'type-safe'-esque workflows limited to classes in in base R, and the package-wide dependencies. Ie. class in, class out. Each method is a soft wrapper of model.frame and therefore the OHLC-V series must be coercible to a data.frame. This rule does not transfer to indicators that uses univariate series, unless passed as a 1-column data.frame or matrix. In such cases, if univariate series is passed as a vector the function calculates the indicator 'as is', and returns a data.frame if the indicator itself is also a univariate series.

The indicator, by default, follows its mathematical definition. However, the cols argument allows for simple rearrangement of the definition by passing relevant columns in a custom order. Refer to the details-section for more on the calculation of the indicators.

Usage

ht_phasor(x, cols, ...)

Arguments

x

An OHLC-V series that is coercible to data.frame. The function assumes that all columns are named in lowercase and order invariant.

cols

An optional formula passed into model.frame. If passed into indicators based on univariate series, the function calculates indicators for each element in 'cols'. For indicators based on multivariate series, it will alter the calculation itself. See vignette("talib") for more details.

...

Additional parameters passed into model.frame

See also

Author

Serkan Korkmaz

Examples

## load Bitcoin (BTC)
## series
data(BTC, package = "talib")

## calculate the indicator
## for Bitcoin (BTC)
output <- talib::ht_phasor(BTC)

## display the results
utils::tail(output)
#>        inphase quadrature
#> 361 -6902.8316   63.92826
#> 362 -5272.2116 3643.07959
#> 363 -3848.5992 4530.11899
#> 364 -2714.7183 5349.44020
#> 365  -198.7199 5615.33995
#> 366  2105.0476  960.98606

## visualize the indicator
## with candlesticks
##
## see ?talib::chart or ?talib::indicator
## for more details
{
 ## chart OHLC-V
 ## series with candlesticks
 talib::chart(BTC)

 ## chart indicator
 ## with default values
 talib::indicator(
     talib::ht_phasor
 )
}