On-Balance Volume

The on_balance_volume() is a generic S3 function that builds upon 'type-safe'-esque workflows limited to classes in in base R, and the package-wide dependencies. Ie. class in, class out. Each method is a soft wrapper of model.frame and therefore the OHLC-V series must be coercible to a data.frame.

Usage

on_balance_volume(x, cols, ...)

Arguments

x

An OHLC-V series that is coercible to data.frame.

cols

(formula). An optional 2 variable formula passed into model.frame. Internally uses ~close + volume by default.

...

Additional parameters passed into model.frame

Value

An object of same class and length of x:

OBV

double

See also

Author

Serkan Korkmaz

Examples

## load Bitcoin (BTC)
## series
data(BTC, package = "talib")

## calculate the indicator
## for Bitcoin (BTC)
output <- talib::on_balance_volume(BTC)

## display the results
utils::tail(output)
#>                          OBV
#> 2024-12-26 01:00:00 38930.56
#> 2024-12-27 01:00:00 35446.60
#> 2024-12-28 01:00:00 36779.98
#> 2024-12-29 01:00:00 34648.52
#> 2024-12-30 01:00:00 30578.67
#> 2024-12-31 01:00:00 33539.63

## visualize the indicator
## with talib::chart()
##
## see ?talib::chart or ?talib::indicator
## for more details
{
 ## chart OHLC-V
 ## series with talib::chart()
 talib::chart(BTC)

 ## chart indicator
 ## with default values
 talib::indicator(
     talib::on_balance_volume
 )
}