Hilbert Transform - Phasor Components

The phasor_components() is a generic S3 function that builds upon 'type-safe'-esque workflows limited to classes in in base R, and the package-wide dependencies. Ie. class in, class out. Each method is a soft wrapper of model.frame and therefore the OHLC-V series must be coercible to a data.frame.

phasor_components() also accepts a double vector in which case the indicator is calculated 'as-is' without passing through model.frame. phasor_components() returns an n by k matrix computed in C by default. When k = 1, the result is simplified to a double vector; for k > 1, the full n by k matrix is returned.

Usage

phasor_components(x, cols, ...)

Arguments

x

An OHLC-V series that is coercible to data.frame. Alternatively, x may also be supplied as a double vector.

cols

(formula). An optional 1 variable formula passed into model.frame. Internally uses ~close by default.

...

Additional parameters passed into model.frame

Value

An object of same class and length of x:

InPhase

double

Quadrature

double

See also

Author

Serkan Korkmaz

Examples

## load Bitcoin (BTC)
## series
data(BTC, package = "talib")

## calculate the indicator
## for Bitcoin (BTC)
output <- talib::phasor_components(BTC)

## display the results
utils::tail(output)
#>                        InPhase Quadrature
#> 2024-12-26 01:00:00 -5284.6163  3630.8079
#> 2024-12-27 01:00:00 -3858.6432  4546.9577
#> 2024-12-28 01:00:00 -2709.2682  5365.0672
#> 2024-12-29 01:00:00  -192.4646  5614.1970
#> 2024-12-30 01:00:00  2110.0585   958.1654
#> 2024-12-31 01:00:00   517.1769 -4075.9341

## visualize the indicator
## with talib::chart()
##
## see ?talib::chart or ?talib::indicator
## for more details
{
 ## chart OHLC-V
 ## series with talib::chart()
 talib::chart(BTC)

 ## chart indicator
 ## with default values
 talib::indicator(
     talib::phasor_components
 )
}