Stochastic Relative Strength Index

The stochastic_relative_strength_index() is a generic S3 function that builds upon 'type-safe'-esque workflows limited to classes in in base R, and the package-wide dependencies. Ie. class in, class out. Each method is a soft wrapper of model.frame and therefore the OHLC-V series must be coercible to a data.frame. This rule does not transfer to indicators that uses univariate series, unless passed as a 1-column data.frame or matrix. In such cases, if univariate series is passed as a vector the function calculates the indicator 'as is', and returns a data.frame if the indicator itself is also a univariate series.

The indicator, by default, follows its mathematical definition. However, the cols argument allows for simple rearrangement of the definition by passing relevant columns in a custom order. Refer to the details-section for more on the calculation of the indicators.

Usage

stochastic_relative_strength_index(
  x,
  cols,
  n = 10,
  n_rsi = 10,
  fast_k = 5,
  fast_d_MAtype = SMA(n = 10),
  ...
)

Arguments

x

An OHLC-V series that is coercible to data.frame. The function assumes that all columns are named in lowercase and order invariant.

cols

An optional formula passed into model.frame. If passed into indicators based on univariate series, the function calculates indicators for each element in 'cols'. For indicators based on multivariate series, it will alter the calculation itself. See vignette("talib") for more details.

n

An integer of length 1.

n_rsi

Time period for relative_strength_index

fast_k

Time period for building the Fast-K line.

fast_d_MAtype

Smoothing for making the Fast-D line.

...

Additional parameters passed into model.frame

See also

Author

Serkan Korkmaz

Examples

## load Bitcoin (BTC)
## series
data(BTC, package = "talib")

## calculate the indicator
## for Bitcoin (BTC)
output <- talib::stochastic_relative_strength_index(BTC)

## display the results
utils::tail(output)
#>         fastk    fastd
#> 361 100.00000 40.72639
#> 362  53.94732 36.12112
#> 363  40.47591 30.16871
#> 364  35.28683 33.69739
#> 365   7.64279 34.46167
#> 366   0.00000 33.92151

## visualize the indicator
## with candlesticks
##
## see ?talib::chart or ?talib::indicator
## for more details
{
 ## chart OHLC-V
 ## series with candlesticks
 talib::chart(BTC)

 ## chart indicator
 ## with default values
 talib::indicator(
     talib::stochastic_relative_strength_index
 )
}