Weighted Close Price

The weighted_close_price() is a generic S3 function that builds upon 'type-safe'-esque workflows limited to classes in in base R, and the package-wide dependencies. Ie. class in, class out. Each method is a soft wrapper of model.frame and therefore the OHLC-V series must be coercible to a data.frame.

Handling of -values

weighted_close_price() iterates over valid values, and returns NA for the remaing part of series.

Usage

weighted_close_price(x, cols, na.ignore = FALSE, ...)

Arguments

x

An OHLC-V series that is coercible to data.frame.

cols

(formula). An optional 3 variable formula passed into model.frame. Internally uses ~high + low + close by default.

na.ignore

(logical). A logical of length 1. FALSE by default. If TRUE 's are ignored during calculation to avoid returning x filled with 's.

...

Additional parameters passed into model.frame

Value

An object of same class and length of x:

WCLPRICE

double

See also

Author

Serkan Korkmaz

Examples

## load Bitcoin (BTC)
## series
data(BTC, package = "talib")

## calculate the indicator
## for Bitcoin (BTC)
output <- talib::weighted_close_price(BTC)

## display the results
utils::tail(output)
#>                     WCLPRICE
#> 2024-12-26 01:00:00 96580.25
#> 2024-12-27 01:00:00 94761.02
#> 2024-12-28 01:00:00 94951.35
#> 2024-12-29 01:00:00 93782.45
#> 2024-12-30 01:00:00 92870.32
#> 2024-12-31 01:00:00 93699.36