Stochastic Relative Strength Index

The stochastic_relative_strength_index() is a generic S3 function that builds upon 'type-safe'-esque workflows limited to classes in in base R, and the package-wide dependencies. Ie. class in, class out. Each method is a soft wrapper of model.frame and therefore the OHLC-V series must be coercible to a data.frame.

Handling of -values

stochastic_relative_strength_index() iterates over valid values, and returns NA for the remaing part of series.

Usage

stochastic_relative_strength_index(
  x,
  cols,
  n = 10,
  n_rsi = 10,
  fastk = 5,
  fastd = SMA(n = 10),
  na.ignore = FALSE,
  ...
)

Arguments

x

An OHLC-V series that is coercible to data.frame.

cols

(formula). An optional 3 variable formula passed into model.frame. Internally uses ~high + low + close by default.

n

(integer). An integer of length 1.

n_rsi

(integer). Period for the relative_strength_index.

fastk

(integer). Period for the fast-k line.

fastd

(list). Period and Moving Average (MA) type for the fast-d line. SMA by default.

na.ignore

(logical). A logical of length 1. FALSE by default. If TRUE 's are ignored during calculation to avoid returning x filled with 's.

...

Additional parameters passed into model.frame

Value

An object of same class and length of x:

FastK

double

FastD

double

See also

Author

Serkan Korkmaz

Examples

## load Bitcoin (BTC)
## series
data(BTC, package = "talib")

## calculate the indicator
## for Bitcoin (BTC)
output <- talib::stochastic_relative_strength_index(BTC)

## display the results
utils::tail(output)
#>                         FastK    FastD
#> 2024-12-26 01:00:00 53.902089 36.11555
#> 2024-12-27 01:00:00 40.406536 30.15621
#> 2024-12-28 01:00:00 35.188699 33.67507
#> 2024-12-29 01:00:00  7.580939 34.43317
#> 2024-12-30 01:00:00  0.000000 33.88678
#> 2024-12-31 01:00:00 88.091719 42.51700

## visualize the indicator
## with talib::chart()
##
## see ?talib::chart or ?talib::indicator
## for more details
{
 ## chart OHLC-V
 ## series with talib::chart()
 talib::chart(BTC)

 ## chart indicator
 ## with default values
 talib::indicator(
     talib::stochastic_relative_strength_index
 )
}